Reworked Prototype to incorporate feedback
Input is now handled by buttons. Each portait is also a button for Specials in the future. Also, the prototype menu now has a button that can be set to True or False. If True, it will play a song that has been pre-programmed for when beats spawn, but where they spawn is random.